jauge double fonctionnelle

+++ ProjetPAC/ProjetPAC/CGauge.h
@@ -17,7 +17,7 @@ namespace chai3d {
 		void update(const double a_frameWidth,
 			const double a_upBoundary,
 			const double a_downBoundary,
-			const double a_gaugesNumber);
+			const int a_gaugesNumber);
 
 	private:
